moved sortQuestsToComplexSingleton to a different class to organize
