+++ src/main/java/aaron/analyzer/user_interface/input/UserInterfaceMain.java
@@ -9,7 +9,6 @@ import aaron.analyzer.user_interface.output.OutputCsv;
 
 import java.io.File;
 import java.io.IOException;
-import java.util.Locale;
 import java.util.Scanner;
 
 public class UserInterfaceMain {
